De-abstractify FrontendCreateVerifyTest
